Updated nanoskimming to a reasonable readable version on the side of jobsubmission
